Move get_version() to config module
