+++ ctms/monitor.py
@@ -1,11 +1,8 @@
 """Application monitoring and health utilities"""
 
-import json
 import logging
-import os.path
 import time
 from datetime import datetime, timezone
-from functools import lru_cache
 
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.sql import func, select
@@ -57,19 +54,3 @@ def check_database(db_session, settings):
     }
 
     return status
-
-
-@lru_cache()
-def get_version():
-    """
-    Return contents of version.json.
-
-    This has generic data in repo, but gets the build details in CI.
-    """
-    ctms_root = os.path.dirname(os.path.dirname(__file__))
-    version_path = os.path.join(ctms_root, "version.json")
-    info = {}
-    if os.path.exists(version_path):
-        with open(version_path, "r", encoding="utf8") as version_file:
-            info = json.load(version_file)
-    return info
